finally got rest of unit tests passing
