+++ test/test_variant_reads_with_dummy_samfile.py
@@ -6,7 +6,7 @@ from varcode import Variant
 from isovar.allele_read import AlleleRead
 from isovar.read_collector import ReadCollector
 
-from mock_read_data import DummySamFile, make_read
+from mock_objects import MockAlignmentFile, make_pysam_read
 from genomes_for_testing import grch38
 
 
@@ -23,11 +23,18 @@ def test_partitioned_read_sequences_snv():
     alt = "G"
 
     variant = Variant(
-        chromosome, location, ref, alt, grch38, normalize_contig_name=False)
+        chromosome,
+        location,
+        ref,
+        alt,
+        grch38,
+        normalize_contig_name=False)
 
-    read = make_read(seq="ACCGTG", cigar="6M", mdtag="3G2")
+    read = make_pysam_read(seq="ACCGTG", cigar="6M", mdtag="3G2")
 
-    samfile = DummySamFile(reads=[read])
+    samfile = MockAlignmentFile(
+        references=(chromosome,),
+        reads=[read])
     read_creator = ReadCollector()
     variant_reads = read_creator.allele_reads_supporting_variant(
         variant=variant,
@@ -58,9 +65,14 @@ def test_partitioned_read_sequences_insertion():
     variant = Variant(
         chromosome, location, ref, alt, grch38, normalize_contig_name=False)
 
-    read = make_read(seq=b"ACCTGTG", cigar="4M1I2M", mdtag="6")
+    read = make_pysam_read(
+        seq=b"ACCTGTG",
+        cigar="4M1I2M",
+        mdtag="6")
 
-    samfile = DummySamFile(reads=[read])
+    samfile = MockAlignmentFile(
+        references=(chromosome,),
+        reads=[read])
     read_creator = ReadCollector()
 
     variant_reads = read_creator.allele_reads_supporting_variant(
@@ -91,9 +103,13 @@ def test_partitioned_read_sequences_deletion():
     variant = Variant(
         chromosome, location, ref, alt, grch38, normalize_contig_name=False)
 
-    read = make_read(seq="ACCTG", cigar="4M1D1M", mdtag="4^T1")
-
-    samfile = DummySamFile(reads=[read])
+    read = make_pysam_read(
+        seq="ACCTG",
+        cigar="4M1D1M",
+        mdtag="4^T1")
+    samfile = MockAlignmentFile(
+        references=(chromosome,),
+        reads=[read])
     read_creator = ReadCollector()
     variant_reads = read_creator.allele_reads_supporting_variant(
         alignment_file=samfile,
